Create pythonclient.py
127.0.0.1 is Localhost IP. It is an IP address that is always defined to be “this machine.”

So, if you’re on a machine that is trying to connect to 127.0.0.1 in effect,
the connection loops back. In other words, the machine tries to go out to
connect to that IP address but instead, it comes back and connects to itself.
That’s why they call it a loop back.

Processing creates a server on this address and the python code connects to it.

It creates a socket object, connects to the server and calls client.send() to send its message. Lastly, it calls client.recv() to read the server’s reply and then prints it.